+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/maven/MojoInit.java
@@ -42,6 +42,9 @@ public class MojoInit extends MojoAdd {
     @Parameter(property = "addFormatting", defaultValue = "false", required = false)
     protected boolean addFormatting;
 
+    @Parameter(property = "withCI", required = false)
+    protected String withCI;
+
     public MojoInit() {
         this.baseInitFolder = ".";
     }
@@ -72,6 +75,7 @@ public class MojoInit extends MojoAdd {
                 context.setBasePackage( this.basePackage );
                 context.setAddJacoco( this.addJacoco );
                 context.setAddFormatting( this.addFormatting );
+                context.setWithCI( this.withCI );
                 this.getLog().info( String.format( "flavour context : %s", context ) );
                 String actualVersion = FlavourFacade.initProject( context );
                 if ( FlavourFacade.FLAVOUR_DIRECT.equals( actualVersion ) ) {