+++ fj-doc-playground-quarkus/src/main/java/org/fugerit/java/doc/playground/init/ProjectInitInput.java
@@ -66,4 +66,8 @@ public class ProjectInitInput {
     @Setter
     private boolean addFormatting;
 
+    @Getter
+    @Setter
+    private String withCI;
+
 }